Refactor : BaseEntity 위치 변경
