+++ src/main/java/umc/kkijuk/server/career/domain/BaseCareer.java
@@ -2,12 +2,9 @@ package umc.kkijuk.server.career.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.kkijuk.server.career.domain.base.BaseEntity;
-import umc.kkijuk.server.careerdetail.domain.CareerDetail;
+import umc.kkijuk.server.common.domian.base.BaseEntity;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
