+++ src/main/java/umc/kkijuk/server/record/domain/Skill.java
@@ -7,9 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import umc.kkijuk.server.career.domain.base.BaseEntity;
-
-import java.time.YearMonth;
+import umc.kkijuk.server.common.domian.base.BaseEntity;
 
 @Entity
 @Getter
